refactor: move public plugin code to api/
