+++ buf.gen.yaml
@@ -4,13 +4,18 @@
 version: v2
 managed:
   enabled: true
+  override:
+    - file_option: go_package_prefix
+      value: github.com/dislogical/cuebe/api/go/proto
 plugins:
   - remote: buf.build/protocolbuffers/go
-    out: pkg
-    opt: paths=source_relative
-  - remote: buf.build/grpc/go:v1.5.1
-    out: pkg
+    out: api/go/proto
+    opt:
+      - paths=source_relative
+      - default_api_level=API_OPAQUE
+  - remote: buf.build/grpc/go
+    out: api/go/proto
     opt:
       - paths=source_relative
 inputs:
-  - directory: pkg
+  - directory: api/proto
