+++ buf.yaml
@@ -4,12 +4,11 @@
 # For details on buf.yaml configuration, visit https://buf.build/docs/configuration/v2/buf-yaml
 version: v2
 modules:
-  - path: pkg
+  - name: buf.build/dislogical/cuebe
+    path: api/proto
 lint:
   use:
     - STANDARD
-  except:
-    - PACKAGE_VERSION_SUFFIX
 breaking:
   use:
     - FILE
